requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-rest-swagger==2.2.0
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
django-heroku==0.3.1
gunicorn==20.0.4
drf-yasg